requirements.txt
streamlit>=1.36
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
PyPDF2>=3.0.1
numpy>=1.26.0
